backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
numpy==1.24.3
scipy==1.11.4
pandas==2.1.4
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
asyncpg==0.29.0
redis[hiredis]==5.0.1
kafka-python==2.0.2
celery==5.3.4
python-dotenv==1.0.0
httpx==0.25.2
